// src/rbm/types.d.ts
export interface Schema$AgentContentMessage {
  /**
   * Information about a file, including the URL of the file and the URL of the file’s thumbnail. The RBM platform serves content from a cache, but an agent can force the RBM platform to fetch a new version of the content and refresh the cache.
   */
  contentInfo?: Schema$ContentInfo;
  /**
   * The unique name of a file. The RBM platform returns a file name when an agent uploads a file.
   */
  fileName?: string | null;
  /**
   * A standalone rich card.
   */
  richCard?: Schema$RichCard;
  /**
   * A list of suggested replies and suggested actions that appear as a list of suggestion chips following the associated agent message. Maximum 11 suggestions. The chips only display when the associated agent message is the most recent message within the conversation (including both agent and user messages). The user can tap a suggested reply to send the text reply back to the agent or tap a suggested action to initiate a native action on the device.
   */
  suggestions?: Schema$Suggestion[];
  /**
   * Text encoded in UTF-8.
   */
  text?: string | null;
}

/**
 * An event from the agent to the user.
 */
export interface Schema$AgentEvent {
  /**
   * The type of the agent event.
   */
  eventType?: string | null;
  /**
   * The ID of the user message that the agent event pertains to. This field is only applicable for agent events of type READ.
   */
  messageId?: string | null;
  /**
   * This field is set by the RBM platform. Do not include it when creating an agent event. The field resolves &quot;phones/{E.164}/agentEvents/{eventId}&quot;, where {E.164} is the user&#39;s phone number in E.164 format and {eventId} is the agent-assigned ID of the agent event.
   */
  name?: string | null;
  /**
   * This field is set by the RBM platform. Do not include it when creating an agent message. The field resolves the time when the event is sent to the user.
   */
  sendTime?: string | null;
}

/**
 * A message sent from the agent to a user.
 */
export interface Schema$AgentMessage {
  /**
   * The content of the agent message.
   */
  contentMessage?: Schema$AgentContentMessage;
  /**
   * This field is set by the RBM platform. Do not include it when creating an agent message. The field resolves &quot;phones/{E.164}/agentMessages/{messageId}&quot;, where {E.164} is the user&#39;s phone number in E.164 format and {messageId} is the agent-assigned ID of the agent message.
   */
  name?: string | null;
  /**
   * This field is set by the RBM platform. Do not include it when creating an agent message. The field resolves the time when the message is sent to the user.
   */
  sendTime?: string | null;
}

/**
 * Request message for RBMService.BatchGetUsers method.
 */
export interface Schema$BatchGetUsersRequest {
  /**
   * List of users&#39; phone numbers in E.164 format.
   */
  users?: string[] | null;
}

/**
 * Response message for RBMService.BatchGetUsers method.
 */
export interface Schema$BatchGetUsersResponse {
  /**
   * Amount of users&#39; phone numbers from the randomly selected list that are RCS-enabled regardless of the launch status for the agent. The ratio between this value and the total_random_sample_user_count can be used to approximate the potential reach for a list of users.
   */
  reachableRandomSampleUserCount?: number | null;
  /**
   * List of users&#39; phone numbers that can be reached from the RBM platform. Only users on carriers that the agent is launched on will be returned.
   */
  reachableUsers?: string[] | null;
  /**
   * Amount of users&#39; phone number randomly selected from the request. Typically this value will be ~75% of the total requested users in the initial BatchGetUsers request. This value will be 0 if the amount of requested users is less than 500.
   */
  totalRandomSampleUserCount?: number | null;
}

/**
 * List of supported features by the user
 */
export interface Schema$Capabilities {
  /**
   * List of RBM features that this device supports.
   */
  features?: string[] | null;
}

/**
 * Card content
 */
export interface Schema$CardContent {
  /**
   * (Optional) Description of the card. Maximum 2000 characters.
   */
  description?: string | null;
  /**
   * (Optional) Media (image, GIF, video) to include in the card.
   */
  media?: Schema$Media;
  /**
   * (Optional) List of suggestions to include in the card. Maximum 4 suggestions.
   */
  suggestions?: Schema$Suggestion[];
  /**
   * (Optional) Title of the card. Maximum 200 characters.
   */
  title?: string | null;
}

/**
 * Carousel of cards.
 */
export interface Schema$CarouselCard {
  /**
   * The list of contents for each card in the carousel. A carousel can have a minimum of 2 cards and a maximum 10 cards.
   */
  cardContents?: Schema$CardContent[];
  /**
   * The width of the cards in the carousel.
   */
  cardWidth?: string | null;
}

/**
 * Message containing the content information.
 */
export interface Schema$ContentInfo {
  /**
   * Publicly reachable URL of the file. The RBM platform determines the MIME type of the file from the content-type field in the HTTP headers when the platform fetches the file. The content-type field must be present and accurate in the HTTP response from the URL.
   */
  fileUrl?: string | null;
  /**
   * If set, the RBM platform fetches the file and thumbnail from the specified URLs, even if the platform has cached copies of the file (and/or of the thumbnail).
   */
  forceRefresh?: boolean | null;
  /**
   * (Optional, for image and video files only) Publicly reachable URL of the thumbnail. If you don&#39;t provide a thumbnail URL, the RBM platform displays a blank placeholder thumbnail until the user&#39;s device downloads the file. Depending on the user&#39;s setting, the file may not download automatically and may require the user to tap a download button.
   */
  thumbnailUrl?: string | null;
}
// ...
export interface Schema$Media {
  /**
   * Information about a file, including the URL of the file and the URL of the file’s thumbnail. The RBM platform serves content from a cache, but an agent can force the RBM platform to fetch a new version of the content and refresh the cache.
   */
  contentInfo?: Schema$ContentInfo;
  /**
   * The unique name of the file, returned by the RBM platform when the file was uploaded.
   */
  fileName?: string | null;
  /**
   * The height of the media within a rich card with a vertical layout. For a standalone card with horizontal layout, height is not customizable, and this field is ignored.
   */
  height?: string | null;
}
// ...
export interface Schema$RichCard {
  /**
   * Carousel of cards.
   */
  carouselCard?: Schema$CarouselCard;
  /**
   * Standalone card.
   */
  standaloneCard?: Schema$StandaloneCard;
}
// ...
export interface Schema$Suggestion {
  /**
   * Users can tap a suggested action to initiate the corresponding native action on the device.
   */
  action?: Schema$SuggestedAction;
  /**
   * Users can tap a suggested reply to send the text reply back to the agent.
   */
  reply?: Schema$SuggestedReply;
}

/**
 * A tester for the agent. The agent can interact with verified testers even if the agent has not yet launched.
 */
export interface Schema$Tester {
  /**
   * The status of the invitation.
   */
  inviteStatus?: string | null;
  /**
   * This field is set by the RBM platform. Do not include it when creating a tester object. The field resolves &quot;phones/{E.164}/testers&quot;, where {E.164} is the tester&#39;s phone number in E.164 format.
   */
  name?: string | null;
}